+++ selfdrive/common/params.cc
@@ -144,6 +144,7 @@ std::unordered_map<std::string, uint32_t> keys = {
     {"DisableUpdates", PERSISTENT},
     {"DisableDisengageOnGas", PERSISTENT},
     {"DisableOnroadUploads", PERSISTENT},
+    {"LowOverheadMode", PERSISTENT},
     {"EnableWideCamera", CLEAR_ON_MANAGER_START},
     {"DoUninstall", CLEAR_ON_MANAGER_START},
     {"DongleId", PERSISTENT},