+++ selfdrive/ui/qt/offroad/settings.cc
@@ -220,6 +220,11 @@ TogglesPanel::TogglesPanel(QWidget *parent) : QWidget(parent) {
                                   "Forces Volt fingerprint",
                                   "../assets/offroad/icon_settings.png",
                                   this));
+  toggles.append(new ParamControl("LowOverheadMode",
+                                  "Lower device overhead",
+                                  "Decreases device power, CPU, and storage use for running better on older hardware by: 1) defaulting to medium brightness (tap DM icon to change), 2) disable onroad logging (loggerd and proclogd). Your device will not keep or upload logs with this enabled!",
+                                  "../assets/offroad/icon_settings.png",
+                                  this));
   toggles.append(new ParamControl("ShowDebugUI",
                                   "Show debug UI elements",
                                   "Show UI elements that aid debugging.",